Simplified version with init files staged from mydotfiles repo instead of declaring explicitly in home-environment record
I tried to use home-gnupg-service-type from rde version of (gnu
home-services gnupg) module. This module has provision to describe
gpg.conf and gpg-agent.conf content through "extra-content" field
which is of type "string-or-gexp". However due to lack mastery of
G-exp, I was unable to provide a gexp for the two configs. So instead
hacked it by reading those files in a string.